requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.3
djangorestframework==3.12.2
gunicorn==20.1.0
joblib==1.0.1
numpy==1.20.3
Pillow==8.2.0
psycopg2==2.8.6
Pygments==2.7.4
pytz==2020.4
scikit-learn==0.24.2
scipy==1.6.3
sqlparse==0.4.1
threadpoolctl==2.1.0
whitenoise==5.2.0
